# docs/swagger.json
{
  "basePath": "/api/v1",
  "definitions": {
    "dto.DeactivateUserResponseDTO": {
      "properties": {
        "deactivated": {
          "example": true,
          "type": "boolean"
        },
        "id": {
          "example": 1,
          "type": "integer"
        }
      },
      "type": "object"
    },
    "dto.RegisterRequest": {
      "properties": {
        "email": {
          "example": "user@example.com",
          "type": "string"
        },
        "full_name": {
          "example": "John Doe",
          "type": "string"
        },
        "password": {
          "example": "strongpassword123",
          "minLength": 6,
          "type": "string"
        }
      },
      "required": [
        "email",
        "full_name",
        "password"
      ],
      "type": "object"
    },
    "dto.UserResponseDTO": {
      "properties": {
        "created_at": {
          "example": "2023-10-27T10:00:00Z",
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "full_name": {
          "type": "string"
        },
        "id": {
          "example": 1,
          "type": "integer"
        }
      },
      "type": "object"
    },
    "update_user.UpdateUserInput": {
      "properties": {
        "email": {
          "type": "string"
        },
        "full_name": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "password": {
          "minLength": 8,
          "type": "string"
        }
      },
      "required": [
        "id"
      ],
      "type": "object"
    }
  },
  "host": "localhost:8080",
  "info": {
    "contact": {
      "name": "API Support"
    },
    "description": "This is a Go backend template using DDD architecture.",
    "title": "Go DDD API Template",
    "version": "1.0"
  },
  "paths": {
    "/users": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Create a new user account",
        "parameters": [
          {
            "description": "User registration details",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.RegisterRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/dto.UserResponseDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/dto.UserResponseDTO"
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "delete": {
        "description": "Deactivate user by their unique ID",
        "parameters": [
          {
            "description": "User ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.DeactivateUserResponseDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/dto.DeactivateUserResponseDTO"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/dto.DeactivateUserResponseDTO"
            }
          }
        },
        "summary": "Deactivate an existing user",
        "tags": [
          "users"
        ]
      },
      "get": {
        "description": "Get user details by their unique ID",
        "parameters": [
          {
            "description": "User ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.UserResponseDTO"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/dto.UserResponseDTO"
            }
          }
        },
        "summary": "Get a user by ID",
        "tags": [
          "users"
        ]
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "description": "Update user details by their unique ID",
        "parameters": [
          {
            "description": "User ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "description": "User update details",
            "in": "body",
            "name": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/update_user.UpdateUserInput"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.UserResponseDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/dto.UserResponseDTO"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/dto.UserResponseDTO"
            }
          }
        },
        "summary": "Update an existing user",
        "tags": [
          "users"
        ]
      }
    }
  },
  "swagger": "2.0"
}